Major V7 update with vertical scrolling. Added a few scumm vars related to scrolling. Dig inventory isn't displayed where it should. I hope it doesn't break anything else in other game...
svn-id: r4066

+++ gfx.h
@@ -106,8 +106,8 @@ struct Gdi {
 	byte _disable_zbuffer;
 
 	bool _useOrDecompress;
-	byte _numLinesToProcess;
-	byte _tempNumLines;
+	int _numLinesToProcess;
+	int _tempNumLines;
 	byte _currentX;
 	byte _hotspot_x;
 	byte _hotspot_y;
@@ -132,7 +132,7 @@ struct Gdi {
 	byte _palette_mod;
 	byte _decomp_shr, _decomp_mask;
 	byte _transparency;
-	uint16 _vertStripNextInc;
+	uint32 _vertStripNextInc;
 	byte *_backupIsWhere;
 
 	/* Bitmap decompressors */
@@ -161,7 +161,7 @@ struct Gdi {
 	void decompressMaskImgOr();
 	void decompressMaskImg();
 
-	void resetBackground(byte top, byte bottom, int strip);
+	void resetBackground(int top, int bottom, int strip);
 	void drawStripToScreen(VirtScreen *vs, int x, int w, int t, int b);
 	void updateDirtyScreen(VirtScreen *vs);
